pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-pydantic-field = "^0.3.9"
httpx = "^0.27.0"
torrentool = "^1.2.0"
python-telegram-bot = "^21.3"
aiokafka = "^0.11.0"
sqlmodel = "^0.0.19"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
aiosqlite = "^0.20.0"
jinja2 = "^3.1.4"
fastapi = "^0.111.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
